chore: 방생성 2번 페이지 폴더화 및 스타일드 컴포넌트 분리
